Change relation between person and bookings to person and booking
