+++ model/Flights.ecore
@@ -40,12 +40,12 @@
   <eClassifiers xsi:type="ecore:EClass" name="Bookings">
     <eStructuralFeatures xsi:type="ecore:EReference" name="bookings" upperBound="-1"
         eType="#//Booking" containment="true"/>
-    <eStructuralFeatures xsi:type="ecore:EReference" name="persons" upperBound="-1"
-        eType="#//Person" eOpposite="#//Person/bookings"/>
   </eClassifiers>
   <eClassifiers xsi:type="ecore:EClass" name="Booking" eSuperTypes="#//FlightObject">
     <eStructuralFeatures xsi:type="ecore:EReference" name="travels" upperBound="-1"
         eType="#//Travel" containment="true"/>
+    <eStructuralFeatures xsi:type="ecore:EReference" name="persons" upperBound="-1"
+        eType="#//Person" eOpposite="#//Person/bookings"/>
   </eClassifiers>
   <eClassifiers xsi:type="ecore:EClass" name="Routes">
     <eStructuralFeatures xsi:type="ecore:EReference" name="routes" upperBound="-1"
@@ -71,7 +71,7 @@
   <eClassifiers xsi:type="ecore:EClass" name="Person" eSuperTypes="#//FlightObject">
     <eStructuralFeatures xsi:type="ecore:EAttribute" name="travelState" eType="#//TravelState"/>
     <eStructuralFeatures xsi:type="ecore:EReference" name="bookings" upperBound="-1"
-        eType="#//Bookings" eOpposite="#//Bookings/persons"/>
+        eType="#//Booking" eOpposite="#//Booking/persons"/>
   </eClassifiers>
   <eClassifiers xsi:type="ecore:EClass" name="Airports">
     <eStructuralFeatures xsi:type="ecore:EReference" name="airports" upperBound="-1"
