+++ src-gen/Flights/Booking.java
@@ -14,6 +14,7 @@ import org.eclipse.emf.common.util.EList;
  * </p>
  * <ul>
  *   <li>{@link Flights.Booking#getTravels <em>Travels</em>}</li>
+ *   <li>{@link Flights.Booking#getPersons <em>Persons</em>}</li>
  * </ul>
  *
  * @see Flights.FlightsPackage#getBooking()
@@ -32,5 +33,19 @@ public interface Booking extends FlightObject {
 	 * @generated
 	 */
 	EList<Travel> getTravels();
+
+	/**
+	 * Returns the value of the '<em><b>Persons</b></em>' reference list.
+	 * The list contents are of type {@link Flights.Person}.
+	 * It is bidirectional and its opposite is '{@link Flights.Person#getBookings <em>Bookings</em>}'.
+	 * <!-- begin-user-doc -->
+	 * <!-- end-user-doc -->
+	 * @return the value of the '<em>Persons</em>' reference list.
+	 * @see Flights.FlightsPackage#getBooking_Persons()
+	 * @see Flights.Person#getBookings
+	 * @model opposite="bookings"
+	 * @generated
+	 */
+	EList<Person> getPersons();
 
 } // Booking
